% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aae.pop}
\alias{aae.pop-package}
\alias{aae.pop}
\title{aae.pop: flexible simulation of multispecies population dynamics}
\description{
aae.pop supports flexible specification of population
  dynamics with tools for simulating single and multispecies dynamics,
  including in metapopulation structures.

  See \code{\link{dynamics}} and \code{\link{simulate}} for detailed
  examples of model definition and simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aae-stats.github.io/aae.pop/}
  \item \url{https://github.com/aae-stats/aae.pop}
  \item Report bugs at \url{https://github.com/aae-stats/aae.pop/issues}
}

}
\keyword{internal}
