% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{covariates}
\alias{covariates}
\title{Specify covariate dependence in models of population dynamics}
\usage{
covariates(masks, funs)
}
\arguments{
\item{masks}{a logical matrix or vector (or list of these)
defining cells affected by \code{funs}. See Details and
\code{\link{masks}}}

\item{funs}{a function or list of functions with one element
for each element of \code{masks}. See Details}
}
\description{
Specify relationship between a vector or matrix of covariates
  and vital rates.
}
\details{
Masks must be of the same dimension as the population
  dynamics matrix and specify cells influenced by covariates
  according to \code{funs}. Functions must take at least
  one arguments, a matrix representing the population
  dynamics matrix. Functions must return a matrix with
  the same dimensions as the input, modified to reflect the
  effects of covariates on vital rates. Additional arguments
  can be passed to \code{funs} and can be specified as
  \code{args} in \code{\link{simulate}}.
}
\examples{
# add
}
