% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_dependence.R
\name{density_dependence}
\alias{density_dependence}
\alias{density_dependence_n}
\title{Specify density dependence in models of population dynamics}
\usage{
density_dependence(masks, funs)

density_dependence_n(masks, funs)
}
\arguments{
\item{masks}{s}

\item{funs}{dfd}
}
\description{
Specify density dependence in vital rates (\code{density_dependence})
  and in total abundances (\code{density_dependence_n}).
}
\details{
something

something
}
\examples{
# add
# add
}
