% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_functions.R
\name{density_functions}
\alias{density_functions}
\alias{beverton_holt}
\alias{ricker}
\title{Common forms of density dependence}
\usage{
beverton_holt(k, exclude = NULL)

ricker(k, exclude = NULL)
}
\arguments{
\item{k}{carrying capacity used to define models of
density dependence. See details for
currently implemented models and their parameters.}

\item{exclude}{vector of classes to exclude from calculation
of total population density. Defaults to NULL, in which
case all classes are used}
}
\description{
Use pre-defined forms of density dependence
  based on common density-dependence functions.
}
\details{
Additional functions are provided to define common
  forms of density dependence. Currently implemented models
  are the Ricker model and Beverton-Holt model, both with
  a single parameter \code{k}.
}
\examples{
# add
# add
}
