% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{dynamics}
\alias{dynamics}
\alias{plot.dynamics}
\title{Create a population dynamics object}
\usage{
dynamics(matrix, ...)

\method{plot}{dynamics}(x, y, ...)
}
\arguments{
\item{matrix}{df}

\item{\dots}{ignored}

\item{x}{an object of class \code{dynamics}}

\item{y}{not supported}
}
\description{
Define population dynamics from a matrix and additional
  objects that determine covariate effects, density dependence, and
  forms of stochasticity.
}
\details{

}
\examples{
# add
}
