% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\alias{get_parameters}
\alias{prior}
\title{Perform inference on a population dynamics model}
\usage{
inference(dynamics, priors, comparison, target, method, ...)

get_parameters(dynamics, args = list())

prior(distribution, ...)
}
\arguments{
\item{dynamics}{a \code{\link{dynamics}} object}

\item{priors}{a list of priors defined with \code{\link{prior}}}

\item{comparison}{a function that takes an output from
\code{\link{simulate}} and formats to match inputs
specified in \code{target}}

\item{target}{values to be compared to simulated model outputs
with \code{comparison}}

\item{method}{a function call for an inference method that
takes as arguments a function and parameters, returning
a vector of summary statistics that define model fit to data.
Example methods include those defined in the \code{EasyABC}
package}

\item{\dots}{additional arguments passed to \code{distribution}}

\item{args}{a list of arguments for ???}

\item{distribution}{xff}
}
\description{
Estimate parameters in a population dynamics model
  defined with the \code{aae.pop} package.
}
\details{
To be completed.

Return parameters from a dynamics object in format
  suitable for inference or for construction of priors.
  Flattened parameters include the model matrix in column-major
  format, followed by any arguments passed to \code{\link{covariates}},
  \code{\link{environmental_stochasticity}},
  \code{\link{demographic_stochasticity}},
  \code{\link{density_dependence}}, and
  \code{\link{density_dependence_n}}, in that order.

PROBABLY BETTER TO DEFINE UPDATEABLE DEFAULTS
  FROM A DYNAMICS OBJECT?
}
\examples{
# add
# add
# add
}
