% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{masks}
\alias{masks}
\alias{reproduction}
\alias{survival}
\alias{transition}
\alias{all_cells}
\alias{all_classes}
\alias{combine}
\alias{combine.mask}
\alias{combine.function}
\alias{combine.default}
\title{Isolate elements of population dynamics models}
\usage{
reproduction(matrix, dims = NULL)

survival(matrix, dims = NULL)

transition(matrix, dims = NULL)

all_cells(matrix, dims = NULL)

all_classes(matrix, dims = NULL)

combine(...)

\method{combine}{mask}(...)

\method{combine}{`function`}(...)

\method{combine}{default}(...)
}
\arguments{
\item{matrix}{a population dynamics matrix for which
a particular mask is required. Only used to determine
mask dimensions, so can be any matrix with appropriate
dimensions}

\item{dims}{a numeric value or vector identifying
subsets of cells to include in a given mask}

\item{\dots}{a set of masks or masking functions
to be combined into a single mask by one of the
\code{combine} methods}
}
\description{
Helper functions to isolate particular components
  of a population dynamics model, such as the reproduction terms,
  transition/growth terms, or particular life stages from an
  abundance vector, such as pre- or post-reproductive stages.
}
\details{
To be completed.

something
}
\examples{
# add
# add
}
