% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.dynamics}
\alias{simulate.dynamics}
\title{Simulate single or multispecies population dynamics in R}
\usage{
\method{simulate}{dynamics}(object, nsim = 1, seed = NULL, init = NULL, options = list(), ...)
}
\arguments{
\item{object}{a \code{dynamics} object created with
\code{\link{dynamics}} or from a subsequent call to
\code{\link{multispecies}} or \code{\link{metapopulation}}}

\item{nsim}{the number of replicate simulations (default = 1)}

\item{seed}{optional seed used prior to initialisation and simulation to
give reproducible results}

\item{init}{an array of initial conditions with one row per replicate and one
column per population stage. Additionally requires one slice per species if
\code{obj} has been created with \code{\link{multispecies}}. Defaults
to \code{NULL}, in which case initial conditions are generated randomly
according to \code{options()$aae.pop_initialisation}}

\item{options}{a named \code{list} of simulation options. Currently accepted
values are:
- \code{ntime} the number of time steps to simulate, ignored if \code{obj}
    includes a \code{\link{covariates}} (default = 50)
- \code{keep_slices} \code{logical} defining whether to keep intermediate
    population abundances or (if \code{FALSE}) to return only the final
    time slice
- \code{tidy_abundances} a function to handle predicted abundance data
    that may be non-integer. Defaults to \code{identity}; suggested
    alternatives are \code{floor}, \code{round}, or \code{ceiling}
- \code{initialise_args} a list of arguments passed to the function
    used to initialise abundance trajectories. Only used if
    \code{init = NULL}. Defaults to \code{options()$aae.pop_lambda},
    which specifies lambda for Poisson random draws. The default
    initialisation function is defined by
    \code{options()$aae.pop_initialisation}.}

\item{\dots}{currently ignored, included for consistency with simulate
generic}
}
\description{
Simulate single or multispecies population dynamics in R
}
\details{
dfd
}
\examples{

# define a population matrix

# define a dynamics objet

# simulate from this
}
