% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate single or multispecies population dynamics in R}
\usage{
simulate(obj, init = NULL, options = list())
}
\arguments{
\item{obj}{a \code{dynamics} object created with
\code{\link{define_dynamics}} or from a subsequent call to
\code{\link{combine_species}} or \code{\link{expand_spatially}}}

\item{init}{an array of initial conditions with one row per replicate and one
column per population stage. Additionally requires one slice per species if
\code{obj} has been created with \code{\link{combine_species}}. Defaults
to \code{NULL}, in which case initial conditions are generated randomly
from a Poisson distribution}

\item{options}{a named \code{list} of simulation options. Currently accepted
values are:
- \code{ntime} the number of time steps to simulate, ignored if \code{obj}
    includes a \code{\link{modifier}} (default = 50)
- \code{replicates} the number of replicate simulations (default = 1000)
- \code{keep_slices} \code{logical} defining whether to keep intermediate
    population abundances or (if \code{FALSE}) to return only the final
    time slice
- \code{tidy_abundances} a function to handle predicted abundance data
    that may be non-integer. Defaults to \code{identity}; suggested
    alternatives are \code{floor}, \code{round}, or \code{ceiling}
- \code{lambda_init} lambda defining a Poisson distribution, used to
    generate rando initial conditions if \code{init = NULL}}
}
\description{
Simulate single or multispecies population dynamics in R
}
\details{

}
\examples{

# define a population matrix

# define a dynamics objet

# simulate from this
}
