% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updaters.R
\name{updaters}
\alias{updaters}
\title{Functions for a single time-step update (matrix %*% vector)}
\description{
Define how population abundances are updated from
  one time step to the next. Functions can take any form but will
  only be vectorised across replicates in limited situations.
}
