% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{covariates}
\alias{covariates}
\title{Specify covariate dependence in models of population dynamics}
\usage{
covariates(x, fun)
}
\arguments{
\item{x}{fk}

\item{fun}{fk}
}
\description{
Specify relationship between a vector or matrix of covariates
  and vital rates.
}
\details{
something
}
\examples{
# add
}
