% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Specify dispersal between populations in a metapopulation model}
\usage{
dispersal(
  kernel,
  stochasticity_masks = NULL,
  stochasticity_funs = NULL,
  density_masks = NULL,
  density_funs = NULL,
  proportion = FALSE
)
}
\arguments{
\item{kernel}{djkfd}

\item{stochasticity_masks}{dfds}

\item{stochasticity_funs}{dfds}

\item{density_masks}{dfds}

\item{density_funs}{dfds}

\item{proportion}{logical indicating whether kernel is specified in absolute
probabilites or as a proportion of the source population (defaults
to \code{FALSE})}
}
\description{
Specify disperal between populations, including stochasticity
  and density dependence in dispersal parameters
}
\details{
something
}
\examples{
# add
}
