% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Specify dispersal between populations in a metapopulation model}
\usage{
dispersal(
  kernel,
  stochasticity_masks = NULL,
  stochasticity_funs = NULL,
  density_masks = NULL,
  density_funs = NULL,
  proportion = FALSE
)
}
\arguments{
\item{kernel}{numeric matrix specifying the probability of
specific classes moving between two populations. Matrices
have the same columns-move-to-rows structure as in
the population dynamics matrices described in
\code{\link{dynamics}}, so a non-zero value in cell (a, b)
denotes a transition from class b in the source population
to class a in the receiving population}

\item{stochasticity_masks}{a logical matrix or list of logical matrices
defining cells affected by \code{stochasticity_funs}.
See Details and \code{\link{masks}}}

\item{stochasticity_funs}{a function or list of functions with
one element for each element of \code{stochasticity_masks}.
See Details}

\item{density_masks}{a logical matrix or list of logical matrices
defining cells affected by \code{density_funs}.
See Details and \code{\link{masks}}}

\item{density_funs}{a function or list of functions with
one element for each element of \code{density_masks}.
See Details}

\item{proportion}{logical indicating whether \code{kernel}
is specified in absolute probabilites or as a proportion
of the source population (defaults to \code{FALSE}).
If \code{TRUE}, values in \code{kernel} are calculated as
a proportion of the total probability an individual
exits that class at any given time step}
}
\description{
Specify disperal between populations, including stochasticity
  and density dependence in dispersal parameters
}
\details{
To be completed.
}
\examples{
# add
}
