% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\name{masks}
\alias{masks}
\alias{reproduction}
\alias{survival}
\alias{transition}
\alias{all_cells}
\alias{all_stages}
\alias{combine}
\alias{combine.masks}
\alias{combine.function}
\alias{combine.default}
\title{Isolate elements of population dynamics models}
\usage{
reproduction(matrix, dims = NULL)

survival(matrix, dims = NULL)

transition(matrix, dims = NULL)

all_cells(matrix, dims = NULL)

all_stages(matrix, dims = NULL)

combine(...)

\method{combine}{masks}(...)

\method{combine}{`function`}(...)

\method{combine}{default}(...)
}
\arguments{
\item{matrix}{s}

\item{dims}{dfd}

\item{\dots}{sj}
}
\description{
Helper functions to isolate particular components
  of a population dynamics model, such as the reproduction terms,
  transition/growth terms, or particular life stages from an
  abundance vector, such as pre- or post-reproductive stages.
}
\details{
something

something
}
\examples{
# add
# add
}
