% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopulation.R
\name{metapopulation}
\alias{metapopulation}
\title{Create a population dynamics object with multiple populations}
\usage{
metapopulation(structure, dynamics, dispersal, ...)
}
\arguments{
\item{structure}{djkfd}

\item{dynamics}{dkjfd}

\item{dispersal}{dkfjd}

\item{\dots}{djfkd}
}
\description{
Define population dynamics for multiple populations of
  a single species linked by dispersal.
}
\details{
djkfdj
}
\examples{
# do stuff
}
