% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies.R
\name{multispecies}
\alias{multispecies}
\alias{pairwise_interaction}
\title{Create a population dynamics object with multiple species}
\usage{
multispecies(interactions, ...)

pairwise_interaction(target, cause, masks, funs)
}
\description{
Define population dynamics for multiple species from
  a set of single-species \code{\link{dynamics}} objects and
  defined pairwise interactions.
}
