% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispecies.R
\name{multispecies}
\alias{multispecies}
\alias{is.multispecies}
\alias{is.interaction}
\title{Create a population dynamics object with multiple species}
\usage{
multispecies(...)

is.multispecies(x)

is.interaction(x)
}
\arguments{
\item{\dots}{\code{pairwise_interaction} objects defining
a set of pairwise interactions between species}
}
\description{
Define population dynamics for multiple species from
  a set of single-species \code{\link{dynamics}} objects and
  defined pairwise interactions.
}
