% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{risk_curve}
\alias{risk_curve}
\title{Calculate (quasi-)extinction risk at multiple thresholds
  for a \code{\link{simulate}} object}
\usage{
risk_curve(sims, threshold = NULL, subset = NULL, times = NULL, n = 100)
}
\arguments{
\item{sims}{an object returned from \code{\link{simulate}}}

\item{threshold}{\code{integer} or \code{numeric} vector
denoting the set of threshold population sizes used to define
the risk curve. Defaults to \code{n} evenly spaced values
from 0 to the maximum observed abundance}

\item{subset}{\code{integer} vector denoting the population classes
to include in calculation of population abundance. Defaults to
all classes}

\item{times}{\code{integer} vector specifying generations to
include in calculation of extinction risk. Defaults to all
simulated generations}

\item{n}{\code{integer} specifying number of threshold values
to use in default case when \code{threshold} is not specified.
Defaults to 100}
}
\description{
Calculate (quasi-)extinction risk at multiple thresholds
  for a \code{\link{simulate}} object
}
\details{
Risk curves represent \code{\link{pr_extinct}} at multiple
  threshold population sizes simultaneously. This gives an expression
  of risk of population declines below a range of values. Risk curves
  are extracted from a \code{\link{simulate}} object as the
  proportion of replicate trajectories that fall below each
  threshold value at any time step within a set period. Abundances
  can be specified for all population classes or for a subset
  of classes.
}
\examples{
# define a basic population
nstage <- 5
popmat <- matrix(0, nrow = nstage, ncol = nstage)
popmat[reproduction(popmat, dims = 4:5)] <- c(10, 20)
popmat[transition(popmat)] <- c(0.25, 0.3, 0.5, 0.65)

# define a dynamics object
dyn <- dynamics(popmat)

# simulate with the default updater
sims <- simulate(dyn)

# simulate with a multinomial updater
sims <- simulate(dyn, options = list(update = update_multinomial))

# calculate risk curve
risk_curve(sims)

# calculate risk curve for multiple thresholds between 0 and 100
risk_curve(sims, threshold = seq(0, 100, by = 1))

# calculate risk curve for 4 and 5 year olds only
pr_extinct(sims, subset = 4:5)

# calculate risk curve but ignore first 10 years
pr_extinct(sims, times = 11:51)
}
