% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{rng}
\alias{rng}
\alias{rmultiunit}
\title{Random number generators not available in existing R
  packages}
\usage{
rmultiunit(
  n,
  mean,
  sd,
  Sigma = NULL,
  Omega = NULL,
  perfect_correlation = FALSE
)
}
\arguments{
\item{y}{ignored; included for consistency with plot method}

\item{class}{integer value specifying which age/stage class to plot.
Defaults to NULL, in which case the sum of all classes is plotted}

\item{\dots}{Additional arguments passed to plot}
}
\description{
Draw random numbers from unusual distributions,
  such as on the unit or non-negative real line with known
  means and standard deviations.
}
\details{
Function to simulate values on unit interval with fixed
  mean, sd, and correlation
}
