% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{rng}
\alias{rng}
\alias{rmultiunit}
\title{Random number generators not available in existing R
  packages}
\usage{
rmultiunit(
  n,
  mean,
  sd,
  Sigma = NULL,
  Omega = NULL,
  perfect_correlation = FALSE
)
}
\arguments{
\item{n}{number of random draws to simulate. Each draw is a vector of
values with length equal to \code{length(mean)} and
\code{length(sd)} and the resulting output has \code{n} rows
and \code{length(mean)} columns}

\item{mean}{vector of mean values on the unit scale}

\item{sd}{vector of positive standard deviations}

\item{Sigma}{optional covariance matrix with dimensions of
\code{length(mean)} by \code{length(mean)} defining covariances
between each pair of values in \code{mean}. Note that only
the correlation structure is retained from \code{Sigma},
so that standard deviations are still required}

\item{Omega}{optional correlation matrix with dimensions of
\code{length(mean)} by \code{length(mean)} defining correlations
between each pair of values in \code{mean}}

\item{perfect_correlation}{\code{logical}, if \code{TRUE}
and \code{Sigma} and \code{Omega} are \code{NULL}, then all
values in each replicate (row) are perfectly correlated with known
mean and standard deviation. If \code{FALSE}, then all values
in each replicate are completely uncorrelated}
}
\description{
Draw random numbers from unusual distributions,
  such as on the unit or non-negative real line with known
  means and standard deviations.
}
\details{
Function to simulate values on unit interval with fixed
  mean, sd, and correlation
}
