% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{rng}
\alias{rng}
\alias{rmultiunit}
\title{Random number generators not available in existing R
  packages}
\usage{
rmultiunit(
  n,
  mean,
  sd,
  Sigma = NULL,
  Omega = NULL,
  perfect_correlation = FALSE
)
}
\arguments{
\item{n}{fd}

\item{mean}{df}

\item{sd}{df}

\item{Sigma}{df}

\item{Omega}{df}

\item{perfect_correlation}{df}
}
\description{
Draw random numbers from unusual distributions,
  such as on the unit or non-negative real line with known
  means and standard deviations.
}
\details{
Function to simulate values on unit interval with fixed
  mean, sd, and correlation
}
