% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate.dynamics}
\alias{is.simulation}
\alias{is.simulation_list}
\alias{print.simulation}
\alias{print.simulation_list}
\alias{plot.simulation}
\alias{plot.simulation_list}
\title{Simulate single or multispecies population dynamics in R}
\usage{
\method{simulate}{dynamics}(
  object,
  nsim = 1,
  seed = NULL,
  init = NULL,
  options = list(),
  args = list()
)

is.simulation(x)

is.simulation_list(x)

\method{print}{simulation}(x, ...)

\method{print}{simulation_list}(x, ...)

\method{plot}{simulation}(x, y, ..., class = NULL)

\method{plot}{simulation_list}(x, y, ..., which = seq_along(x))
}
\arguments{
\item{object}{a \code{dynamics} object created with
\code{\link{dynamics}} or from a subsequent call to
\code{\link{multispecies}} or \code{\link{metapopulation}}}

\item{nsim}{the number of replicate simulations (default = 1)}

\item{seed}{optional seed used prior to initialisation and simulation to
give reproducible results}

\item{init}{an array of initial conditions with one row per replicate and one
column per population stage. Additionally requires one slice per species if
\code{obj} has been created with \code{\link{multispecies}}. Defaults
to \code{NULL}, in which case initial conditions are generated randomly
according to \code{options()$aae.pop_initialisation}}

\item{options}{a named \code{list} of simulation options. Currently accepted
values are:
- \code{ntime} the number of time steps to simulate, ignored if \code{obj}
    includes a \code{\link{covariates}} (default = 50)
- \code{keep_slices} \code{logical} defining whether to keep intermediate
    population abundances or (if \code{FALSE}) to return only the final
    time slice
- \code{tidy_abundances} a function to handle predicted abundance data
    that may be non-integer. Defaults to \code{identity}; suggested
    alternatives are \code{floor}, \code{round}, or \code{ceiling}
- \code{initialise_args} a list of arguments passed to the function
    used to initialise abundance trajectories. Only used if
    \code{init = NULL}. Defaults to \code{options()$aae.pop_lambda},
    which specifies lambda for Poisson random draws. The default
    initialisation function is defined by
    \code{options()$aae.pop_initialisation}.}

\item{args}{named list of lists passing arguments to processes defined
in \code{object}, including \code{interaction} for
\code{\link{multispecies}} objects.}

\item{x}{a simulation or simulation_list object}

\item{\dots}{Additional arguments passed to plot}

\item{y}{ignored; included for consistency with plot method}

\item{class}{integer value specifying which age/stage class to plot.
Defaults to NULL, in which case the sum of all classes is plotted}

\item{which}{integer value or vector of integer values specifying
which species to plot. Defaults to one plot for each species}
}
\description{
Simulate population dynamics for one or more
  species defined by \code{\link{dynamics}} objects.
}
\details{
To be completed.

Basic plot methods are provided for \code{simulation} and
  \code{simulation_list} classes. Plots can present a single class
  or the sum of all classes, and can include one or more species
  for \code{simulation_list} objects. \code{\dots} can be used to
  pass additional arguments to \code{\link[graphics]{plot}} and
  \code{\link[graphics]{lines}}.
}
\examples{

# define a population matrix

# define a dynamics objet

# simulate from this
}
