% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasticity.R
\name{stochasticity}
\alias{stochasticity}
\alias{environmental_stochasticity}
\alias{demographic_stochasticity}
\title{Specify environmental and demographic stochasticity in models of
  population dynamics}
\usage{
environmental_stochasticity(masks, funs)

demographic_stochasticity(masks, funs)
}
\arguments{
\item{masks}{dfds}

\item{funs}{d}
}
\description{
Specify environmental stochasticity (random variation
  in vital rates) and demographic stochasticity (random variation
  in population outcomes).
}
\details{
something

something
}
\examples{
# add
# add
}
