% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{templates}
\alias{templates}
\alias{get_template}
\alias{murraycod}
\title{Use a fully parameterised population dynamics object}
\usage{
get_template(sp, x = NULL, ...)

murraycod(x = NULL, ...)
}
\arguments{
\item{sp}{string specifying species common name (see details for
a list of species currently included)}

\item{x}{a vector or matrix of covariates used to define
time-varying vital rates according to a pre-defined
functional response}

\item{\dots}{additional objects passed to \code{\link{dynamics}},
overwriting the default, pre-defined values for these classes.
Must be one or more of \code{\link{covariates}},
\code{\link{environmental_stochasticity}},
\code{\link{demographic_stochasticity}},
\code{\link{density_dependence}}, or
\code{\link{density_dependence_n}}}
}
\description{
Use pre-defined population dynamics objects to
  define a matrix model for a species with known parameters,
  optionally including new covariates.
}
\details{
The \code{get_template} function and associated
  wrapper functions (e.g. \code{murraycod}) return a collated
  \code{\link{dynamics}} object parameterised with values based
  on existing data sets and published works. Currently implemented
  species are: Murray cod (*Maccullochella peelii*),
  trout cod (*Maccullochella macquariensis*), golden perch
  (*Macquaria ambigua*), silver perch (*Bidyanus bidyanus*),
  Macquarie perch (*Macquaria australasica*), Australian smelt
  (*Retropinna semonia*), and common galaxias (*Galaxias maculatus*).
}
\examples{
# add
}
