% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{templates}
\alias{templates}
\alias{murray_cod}
\alias{macquarie_perch}
\alias{get_args}
\title{Parameterised population dynamics objects}
\usage{
murray_cod(k = 20000, ...)

macquarie_perch(k = 1000, ...)

get_args(sp, ...)
}
\arguments{
\item{k}{carrying capacity}

\item{\dots}{additional arguments passed to templates or args
functions}
}
\description{
Use pre-defined population dynamics objects to
  define a matrix model for a species with known parameters.
}
\details{
These functions (e.g. \code{murray_cod}) return a collated
  \code{\link{dynamics}} object parameterised with values based
  on existing data sets and published works. Currently implemented
  species are: Murray cod (*Maccullochella peelii*) and Macquarie
  perch (*Macquaria australasica*).

  In some cases, additional arguments might need to be passed to
  \code{\link{simulate}}, and it may be useful to define these as
  part of the template. The \code{get_args} function
  handles this situation. Arguments should be
  specified with \code{args_} followed by
  a species name or identifier (e.g \code{args_my_species}).
  Arguments functions should return a series of named lists
  for any of \code{args}, \code{args.dyn}, or \code{args.fn} (see
  \code{\link{simulate}} for descriptions of these terms).
}
\examples{
# define a basic model for Murray cod with
#   carrying capacity = 25000
mc <- murray_cod(k = 25000)

# simulate from this model
sims <- simulate(mc, nsim = 100)

# plot the simulated values
plot(sims)
}
