% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updaters.R
\name{updaters}
\alias{updaters}
\alias{update_crossprod}
\alias{update_binomial_leslie}
\alias{update_multinomial}
\title{Functions for a single time-step update (matrix %*% vector)}
\usage{
update_crossprod(pop, mat)

update_binomial_leslie(pop, mat)

update_multinomial(pop, mat)
}
\arguments{
\item{pop}{current state of the population}

\item{mat}{matrix of vital rates used to update population state}
}
\description{
Define how population abundances are updated from
  one time step to the next. Functions can take any form but will
  only be vectorised across replicates in limited situations.
}
\details{
Update abundances with a direct matrix multiplication
  that does not include any form of demographic stochasticity.
  This is the fastest update option and will vectorise across
  replicates if the population matrix is not expanded by
  \code{\link{environmental_stochasticity}} or
  \code{\link{density_dependence}}.

Update abundances with a direct RNG draw

Update abundances with a direct RNG draw
  that combines update with demographic stochasticity,
  allowing for general matrix forms (slower than
  \code{update_binomial_leslie}).
}
